refactor out child classes, in aim of avoiding serialisation error
